pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-c2n-repro"
version = "0.1.0"
description = "Reproducible C2N fraud detection pipelines"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.23",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "xgboost>=2.0",
  "category-encoders>=2.6",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "torch>=2.0",
  "networkx>=3.0",
  "catboost>=1.2",
  "imbalanced-learn>=0.11",
  "pyod>=1.1",
  "pytorch-tabnet>=4.1",
]

[tool.setuptools.packages.find]
where = ["src"]
